services/redirect/requirements.txt
fastapi
uvicorn[standard]
boto3
psycopg2-binary